+++ graphql/index.js
@@ -4,6 +4,6 @@ const { graphql: scalars } = require('@orbiting/backend-modules-scalars')
 const local = addTypes(loadModule(__dirname), [scalars])
 
 const { graphql: auth } = require('@orbiting/backend-modules-auth')
-const crowdfundings = require('./modules/crowdfundings/graphql')
+const crowdfundings = require('../modules/crowdfundings/graphql')
 
 module.exports = merge(local, [crowdfundings, auth])
