+++ modules/crowdfundings/lib/Pledge.js
