+++ modules/crowdfundings/lib/payments/postfinance/sha.js
