+++ modules/crowdfundings/lib/payments/sendPaymentSuccessful.js
