+++ modules/crowdfundings/lib/payments/stripe/addSource.js
