+++ modules/crowdfundings/lib/sendPendingPledgeConfirmations.js
